+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 7743b96 - v8
+| Author: FJrodafo - Tue, 30 Jul 2024 18:04:00 +0100
+| 
 * Commit: 87cf90f - v7
 | Author: FJrodafo - Sat, 27 Jul 2024 13:55:57 +0100
 |